+++ notify-debouncer-full/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## unreleased
 
 - PERF: park the debouncer thread to avoid idle polling [#933]
+- CHANGE: `FileIdCache::rescan` takes `&[WatchRoot]` instead of `&[(PathBuf, RecursiveMode)]` **breaking**
 
 [#933]: https://github.com/notify-rs/notify/pull/933